perception: finish debugging offline_sequential_obstacle_perception_test

+++ modules/perception/obstacle/common/file_system_util.cc
@@ -36,7 +36,7 @@ std::string GetFileName(const std::string& path) {
 
 void GetFileNamesInFolderById(const std::string& folder, const std::string& ext,
                               std::vector<std::string>* ret) {
-  std::vector<int> ret_id;
+  std::vector<double> ret_id;
   ret->clear();
   namespace fs = boost::filesystem;
   if (!fs::exists(folder) || !fs::is_directory(folder)) {
@@ -53,7 +53,7 @@ void GetFileNamesInFolderById(const std::string& folder, const std::string& ext,
       std::string temp_id_str =
           temp_path.substr(temp_path.rfind('_') + 1,
                            temp_path.rfind('.') - temp_path.rfind('_') - 1);
-      int temp_id = std::atoi(temp_id_str.c_str());
+      double temp_id = std::atof(temp_id_str.c_str());
       ret_id.push_back(temp_id);
     }
     ++it;
@@ -63,7 +63,7 @@ void GetFileNamesInFolderById(const std::string& folder, const std::string& ext,
   for (int i = 0; i < ret_size; ++i) {
     for (int j = i; j < ret_size; ++j) {
       if (ret_id[i] > ret_id[j]) {
-        int temp_id = ret_id[i];
+        double temp_id = ret_id[i];
         ret_id[i] = ret_id[j];
         ret_id[j] = temp_id;
         std::string temp_path = (*ret)[i];
